Fixed uint overflow.

+++ TightWiki.Security/Helpers.cs
@@ -23,11 +23,11 @@ namespace TightWiki.Security
             return result.ToUpper();
         }
 
-        public static uint Crc32(string text)
-            => (new Crc32()).Get(Encoding.Unicode.GetBytes(text));
+        public static int Crc32(string text)
+            => (int)(new Crc32()).Get(Encoding.Unicode.GetBytes(text));
 
-        public static uint Crc32(byte[] bytes)
-            => (new Crc32()).Get(bytes);
+        public static int Crc32(byte[] bytes)
+            => (int)(new Crc32()).Get(bytes);
 
         public static string Sha1(string text)
             => BitConverter.ToString(SHA1.HashData(Encoding.Unicode.GetBytes(text))).Replace("-", "");
